refactor: 회원 엔티티 매핑 개선
